Move up some request attrs in class hierarchy.
Move up some attributes from ModelExecRequest to DmodJobRequest.
